refactor: Update ProducerAndConsumer class in Program.cs
This modification changes the ProducerAndConsumer class in Program.cs. It moves the initialization of `_bus` and `_id` to the class level from the constructor. It also modifies the console output for publishing and receiving messages by adding the `_id` value in the logged string, and changes the 'HandleTextMessage' method from static to non-static.